project 4 refactor, fix MultiThreadRenderer
